proper impl and bug fixes
