+++ djs-bot/commands/music/previous.js
@@ -36,7 +36,8 @@ const command = new SlashCommand()
 		});
 	}
 
-	const {status, previousSong} = playerUtil.playPrevious(player);
+	const previousSong = player.queue.previous;
+	const status = await playerUtil.playPrevious(player);
 
 	if (status === 1) return interaction.reply({
 		embeds: [
