+++ djs-bot/interactions/next.js
@@ -1,5 +1,7 @@
 const SlashCommand = require("../lib/SlashCommand");
 const { ccInteractionHook } = require("../util/interactions");
+const playerUtil = require("../util/player");
+const { redEmbed } = require("../util/embeds");
 
 const command = new SlashCommand()
 	.setName("next")
@@ -12,8 +14,18 @@ const command = new SlashCommand()
 
 		const { player, channel, sendError } = data;
 
-		if (player.playing) {
-			player.stop();
+		const song = player.queue.current;
+		const status = playerUtil.skip(player);
+
+		if (status === 1) {
+			return interaction.reply({
+				embeds: [
+					redEmbed(
+						`There is nothing after [${song.title}](${song.uri}) in the queue.`
+					),
+				],
+				ephemeral: true,
+			});
 		}
 
 		return interaction.deferUpdate();
